# drivebook-hybrid/openapi.yaml
swagger: '2.0'
info:
  title: DriveBook Voice Service API
  description: API for AI voice agent to interact with DriveBook platform for booking driving lessons
  version: 1.0.0
  contact:
    name: DriveBook Support
host: your-service-url.railway.app
basePath: /api
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json

paths:
  /health:
    get:
      summary: Health check endpoint
      description: Check if the voice service is running and database is connected
      operationId: healthCheck
      tags:
        - System
      responses:
        '200':
          description: Service is healthy
          schema:
            type: object
            properties:
              status:
                type: string
                example: ok
              uptime:
                type: number
                example: 123.45
              database:
                type: string
                example: connected
              timestamp:
                type: string
                format: date-time

  /instructor/lookup:
    get:
      summary: Find instructor by phone number
      description: Look up instructor details using their phone number to check availability and details
      operationId: lookupInstructor
      tags:
        - Instructors
      parameters:
        - name: phone
          in: query
          type: string
          required: true
          description: Instructor's phone number in international format (e.g., +61412345678)
          pattern: '^\+[1-9]\d{1,14}$'
      responses:
        '200':
          description: Instructor found successfully
          schema:
            type: object
            properties:
              id:
                type: string
                description: Unique instructor identifier
              name:
                type: string
                description: Instructor's full name
              phone:
                type: string
                description: Instructor's phone number
              email:
                type: string
                description: Instructor's email address
              available:
                type: boolean
                description: Whether instructor is currently accepting bookings
        '400':
          description: Invalid phone number format
          schema:
            type: object
            properties:
              error:
                type: string
        '404':
          description: Instructor not found
          schema:
            type: object
            properties:
              error:
                type: string
        '500':
          description: Server error
          schema:
            type: object
            properties:
              error:
                type: string

  /bookings:
    post:
      summary: Create a new booking
      description: Book a driving lesson with an instructor for a specific date and time
      operationId: createBooking
      tags:
        - Bookings
      parameters:
        - name: body
          in: body
          required: true
          description: Booking details
          schema:
            type: object
            required:
              - instructorId
              - clientPhone
              - date
              - time
            properties:
              instructorId:
                type: string
                description: Unique identifier of the instructor
                example: "clx123abc456"
              clientPhone:
                type: string
                description: Client's phone number in international format
                example: "+61412345678"
                pattern: '^\+[1-9]\d{1,14}$'
              clientName:
                type: string
                description: Client's full name (optional)
                example: "John Smith"
              date:
                type: string
                format: date
                description: Lesson date in YYYY-MM-DD format
                example: "2026-03-15"
              time:
                type: string
                description: Lesson time in HH:MM format (24-hour)
                example: "14:00"
              duration:
                type: integer
                description: Lesson duration in minutes
                example: 60
                default: 60
              notes:
                type: string
                description: Additional notes or special requests
                example: "First lesson, needs pickup from home"
      responses:
        '201':
          description: Booking created successfully
          schema:
            type: object
            properties:
              bookingId:
                type: string
                description: Unique booking identifier
                example: "bk_abc123xyz"
              message:
                type: string
                example: "Booking confirmed"
              smsConfirmation:
                type: boolean
                description: Whether SMS confirmation was sent
        '400':
          description: Invalid booking data
          schema:
            type: object
            properties:
              error:
                type: string
        '409':
          description: Time slot not available
          schema:
            type: object
            properties:
              error:
                type: string
        '500':
          description: Server error
          schema:
            type: object
            properties:
              error:
                type: string

  /voice/incoming:
    post:
      summary: Handle incoming voice call
      description: Twilio webhook endpoint for incoming calls - handles voice interaction
      operationId: handleIncomingCall
      tags:
        - Voice
      parameters:
        - name: body
          in: body
          required: true
          description: Twilio webhook payload
          schema:
            type: object
            properties:
              From:
                type: string
                description: Caller's phone number
              To:
                type: string
                description: Called phone number
              CallSid:
                type: string
                description: Unique call identifier
      responses:
        '200':
          description: TwiML response
          schema:
            type: string
            example: '<?xml version="1.0" encoding="UTF-8"?><Response>...</Response>'
        '403':
          description: Invalid Twilio signature
          schema:
            type: object
            properties:
              error:
                type: string

  /voice/voicemail:
    post:
      summary: Handle voicemail recording
      description: Twilio webhook endpoint for voicemail status callbacks
      operationId: handleVoicemail
      tags:
        - Voice
      parameters:
        - name: body
          in: body
          required: true
          description: Twilio recording callback payload
          schema:
            type: object
            properties:
              RecordingUrl:
                type: string
                description: URL of the recorded voicemail
              RecordingSid:
                type: string
                description: Unique recording identifier
              From:
                type: string
                description: Caller's phone number
      responses:
        '200':
          description: Voicemail processed
          schema:
            type: object
            properties:
              status:
                type: string
                example: saved

definitions:
  Error:
    type: object
    properties:
      error:
        type: string
        description: Error message
      requestId:
        type: string
        description: Request identifier for debugging

  Booking:
    type: object
    properties:
      id:
        type: string
      instructorId:
        type: string
      clientPhone:
        type: string
      date:
        type: string
        format: date
      time:
        type: string
      duration:
        type: integer
      status:
        type: string
        enum: [pending, confirmed, completed, cancelled]
      createdAt:
        type: string
        format: date-time

  Instructor:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      phone:
        type: string
      email:
        type: string
      available:
        type: boolean
      rating:
        type: number
        format: float
      totalLessons:
        type: integer

securityDefinitions:
  ApiKeyAuth:
    type: apiKey
    in: header
    name: X-API-Key
    description: API key for authentication (optional for development)

tags:
  - name: System
    description: System health and status endpoints
  - name: Instructors
    description: Instructor lookup and management
  - name: Bookings
    description: Booking creation and management
  - name: Voice
    description: Twilio voice webhook endpoints